Add a unit test test/Test.hs
The test-cases are based on examples collected from hackage, HOOD webpage, and
from 2000 SIGPLAN paper.

Test harness:
(caughtO FORM) runs (printO FORM) and captures
  the value written to stdout and the observation written to stderr.
(tc FORM EXPECTED-VALUE EXPECTED-OBS) makes a HUnit TestCase

Status: the tests pass both compiled and in repl.
That is, these both work: 'stack test' and 'stack repl --test; :main'

In the repl, a function can be observed as a set of maps, instead of as one map.
To tolerate this, added a second test-case generator
  tc2::Show a => a -> Value -> [Observation] -> Test